requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
google-auth==2.40.2
google-auth-oauthlib==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
redis==6.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
spotipy==2.25.1
urllib3==2.4.0
Werkzeug==3.1.3
ytmusicapi==1.10.3
